requirements.txt
spacy==3.7.2
transformers==4.36.2
torch==2.1.2
scikit-learn==1.3.2
streamlit==1.29.0
google-generativeai==0.3.2
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0 